srcs/User-management/User_service/requirements.txt
Django==4.2
djangorestframework
djangorestframework-simplejwt
gunicorn
psycopg2-binary
django-environ
python-dotenv
requests